+++ unhaunter/src/app.rs
@@ -24,7 +24,7 @@ use unplayer::plugin::UnhaunterPlayerPlugin;
 use unprofile::plugin::UnhaunterProfilePlugin;
 use unsettings::plugin::UnhaunterSettingsPlugin;
 use unstd::materials::{CustomMaterial1, UIPanelMaterial};
-use unstd::picking::CustomSpritePickingPlugin;
+use unstd::picking::{CustomSpritePickingPlugin, GamepadPointerPlugin};
 use unstd::plugins::board::UnhaunterBoardPlugin;
 use unstd::plugins::manual::UnhaunterManualPlugin;
 use unstd::plugins::root::UnhaunterRootPlugin;
@@ -68,7 +68,7 @@ pub fn app_run(cli_options: CliOptions) {
         .add_plugins(UiMaterialPlugin::<UIPanelMaterial>::default());
 
     // Add picking support for our custom sprites
-    app.add_plugins(CustomSpritePickingPlugin);
+    app.add_plugins((CustomSpritePickingPlugin, GamepadPointerPlugin));
 
     app.add_plugins((
         UnhaunterCorePlugin,